requirements.txt
# Excel 工具箱依赖包
# Core dependencies - 轻量级Excel处理方案
xlrd>=2.0.0  # 支持读取旧版Excel文件(.xls)
xlsxwriter>=3.1.0  # 写入Excel文件，不依赖numpy
Pillow>=10.0.0

# 轻量级xlsx读取（替代openpyxl）
defusedxml>=0.7.0  # 安全的XML解析
lxml>=4.9.0  # 快速XML处理

# PDF processing dependencies
pypdf>=3.17.0
pdf2image>=1.16.3
pytesseract>=0.3.10
reportlab>=4.0.0

# PaddleOCR dependencies (full framework, ~600MB, best accuracy)
# 支持 2.x 和 3.x 版本
paddlepaddle>=2.5.0  # PaddlePaddle deep learning framework (CPU version)
# paddlepaddle-gpu>=2.5.0  # GPU版本（如需GPU加速，取消注释此行并注释上一行）
paddleocr>=2.7.0     # PaddleOCR library (2.x or 3.x)

# RapidOCR (轻量级替代方案, ~70MB, 推荐用于打包)
# 使用 PaddleOCR 相同的模型，但通过 ONNX Runtime 推理，打包体积大幅减小
rapidocr-onnxruntime>=1.3.0

# Database dependencies (optional, for template/config sharing)
sqlalchemy>=2.0.0  # Database ORM
pymysql>=1.1.0  # MySQL driver (optional)
psycopg2-binary>=2.9.0  # PostgreSQL driver (optional)
# SQLite is built-in, no extra package needed

# Build dependencies
pyinstaller>=6.0.0

# Testing dependencies (development only)
pytest>=7.4.0
pytest-cov>=4.1.0

# PPT to PDF conversion
comtypes>=1.2.0

# Optional: for better development experience
# typing-extensions>=4.8.0  # Python < 3.10 需要
